% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ankle_angle.R
\name{ankle_angle}
\alias{ankle_angle}
\title{ankle_angle}
\usage{
ankle_angle(data_array, stand = TRUE)
}
\arguments{
\item{data_array}{data array.}

\item{stand}{bool.}
}
\description{
ankle_angle can estimate ankle angle based on some.
}
