% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_detect.R
\name{dif_detect}
\alias{dif_detect}
\title{dif_detect}
\usage{
dif_detect(data_array, width = 15, scaling = 50)
}
\arguments{
\item{data_array}{data array.}

\item{width}{width is a parameter.}

\item{scaling}{scaling is a scaling parameter.}
}
\description{
dif_detect() can detect anomaly data.
}
