% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_group.R
\name{epoch_group}
\alias{epoch_group}
\title{epoch_group}
\usage{
epoch_group(List, group)
}
\arguments{
\item{List}{it is epoch_list.}

\item{group}{grouping vector}
}
\description{
epoch_group() split epoch_list to groups.
}
