% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_plot.R
\name{epoch_plot}
\alias{epoch_plot}
\title{epoch_plot}
\usage{
epoch_plot(
  epoch_list,
  Left = TRUE,
  file_name = NULL,
  joint_name = NULL,
  xyplot = "xy"
)
}
\arguments{
\item{epoch_list}{epoch_list.}

\item{Left}{bool.}

\item{file_name}{file name.}

\item{joint_name}{joint name.}

\item{xyplot}{xy plot.}
}
\description{
plot for joints' coordinates data after epoching.
}
