% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hip_angle.R
\name{hip_angle}
\alias{hip_angle}
\title{hip_angle}
\usage{
hip_angle(data_array, stand = TRUE)
}
\arguments{
\item{data_array}{data array.}

\item{stand}{bool.}
}
\description{
hip_angle() estimates hip angles based on some.
}
