% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmp_imp.R
\name{jmp_imp}
\alias{jmp_imp}
\title{jmp_imp}
\usage{
jmp_imp(epoch_list, lag = 5, percentile = 95, except_joint = NULL)
}
\arguments{
\item{epoch_list}{data array.}

\item{lag}{lag.}

\item{percentile}{percentile.}

\item{except_joint}{vector of except joints}
}
\description{
jmp_imp() can impute missing data.
}
