% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_plot.R
\name{joint_plot}
\alias{joint_plot}
\title{plot for joint's coordinates}
\usage{
joint_plot(data_array, file_name = NULL, joint_name = NULL, xyplot = "xy")
}
\arguments{
\item{data_array}{data array (it is obtained from function make_array()).}

\item{file_name}{plot title.}

\item{xyplot}{plot xy-coord x-coord, y-coord.}

\item{joiint_name}{a joint name/}
}
\description{
plotting xy-coordinates of a joint.
}
