% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_imp.R
\name{jump_imp}
\alias{jump_imp}
\title{jump_imp}
\usage{
jump_imp(epoch_list, sd_ratio = 1, lag = 3, rept = 5, except_joint = NULL)
}
\arguments{
\item{epoch_list}{data array.}

\item{sd_ratio}{sd*sd_ratio}

\item{lag}{lag.}

\item{rept}{the number of repeat.}

\item{except_joint}{vector of except joints}
}
\description{
jump_imp() can impute missing data.
}
