% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_array.R
\name{make_array}
\alias{make_array}
\title{making array from OpenPose JSON files}
\usage{
make_array(filespath = NULL, thr = 0)
}
\arguments{
\item{filespath}{a directory path containing OpenPose JSON files.}

\item{thr}{a threshold for OpenPose score data.}
}
\description{
making array from OpenPose JSON files.
}
