% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansd.R
\name{meansd}
\alias{meansd}
\title{meansd}
\usage{
meansd(epoch_list, Jlist)
}
\arguments{
\item{epoch_list}{epoch list.}

\item{Jlist}{joint list.}
}
\description{
meansd() can convert epoch list to data array.
}
