% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_anno.R
\name{median_anno}
\alias{median_anno}
\title{median_anno}
\usage{
median_anno(epoch_list, Jlist)
}
\arguments{
\item{epoch_list}{data array.}

\item{except}{vector of except joints}
}
\description{
sw_detect() is a function for detecting misrecognition of left and right.
}
