% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_imp.R
\name{miss_imp}
\alias{miss_imp}
\title{miss_imp}
\usage{
miss_imp(data_array, thr = 0)
}
\arguments{
\item{data_array}{data array.}

\item{thr}{the threshold parameter about confident score.}
}
\description{
miss_imp() can detect missing data.
}
