% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_diff.R
\name{NA_diff}
\alias{NA_diff}
\title{NA_diff}
\usage{
NA_diff(data_array, lag = 5, percentile = 95)
}
\arguments{
\item{data_array}{data array.}

\item{lag}{lag parameter (used in diff()).}

\item{percentile}{percentile.}
}
\description{
NA_diff convert coordinates to NA based on diff().
}
