% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_score.R
\name{NA_score}
\alias{NA_score}
\title{NA_score}
\usage{
NA_score(data_array, thr = 0)
}
\arguments{
\item{data_array}{data array.}

\item{thr}{threshold.}
}
\description{
NA_score can convert coordinates to NA based on OpenPose score.
}
