% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{overlap}
\usage{
overlap(data_array, Jlist)
}
\arguments{
\item{data_array}{data array.}

\item{lag}{lag.}

\item{thr}{threshold}

\item{fit}{fitting threshold}
}
\description{
overlap() can detect overlapping foot.
}
