% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_det.R
\name{peak_det}
\alias{peak_det}
\title{peak_det}
\usage{
peak_det(y, w = 1, span = 0.05)
}
\arguments{
\item{y}{y is a coordinate time series.}

\item{w}{w is tuning parameter.}

\item{span}{span is tuning parameter.}
}
\description{
peak_det() can detect peaks from coordinates data.
}
