% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resampling}
\alias{resampling}
\title{resampling}
\usage{
resampling(TS, to = NULL)
}
\arguments{
\item{TS}{times series vector.}

\item{to}{arbitrary length}
}
\description{
convert vector to arbitrary length vector
}
