% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{standardize}
\usage{
standardize(data_array, centering = TRUE, scaling = TRUE, anchor = "Neck")
}
\arguments{
\item{data_array}{input data array.}

\item{centering}{bool variable, center coordinate or not}

\item{scaling}{bool variable, center coordinate or not}
}
\description{
standardize() can standardize data array.
}
