% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_detect.R
\name{sw_detect}
\alias{sw_detect}
\title{sw_detect}
\usage{
sw_detect(epoch_list, except)
}
\arguments{
\item{epoch_list}{data array.}

\item{except}{vector of except joints}
}
\description{
sw_detect() is a function for detecting misrecognition of left and right.
}
