% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switching.R
\name{switching}
\alias{switching}
\title{switching}
\usage{
switching(epoch_list, lag = 5, thr = 0.3, fit = 0.3)
}
\arguments{
\item{epoch_list}{epoch list.}

\item{lag}{lag.}

\item{thr}{threshold}

\item{fit}{fitting threshold}
}
\description{
switching() can detect NA and impute them.
}
