% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{trajectory}
\alias{trajectory}
\title{trajectory}
\usage{
trajectory(Data, joint_name = NULL, file_name = NULL)
}
\arguments{
\item{Data}{Data.}

\item{joint_name}{joint name.}

\item{file_name}{file name.}
}
\description{
trajectory() can plot coordinates for trajectory analysis.
}
