% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_color.R
\name{trajectory_color}
\alias{trajectory_color}
\title{trajectory_color}
\usage{
trajectory_color(
  coord_data,
  emg_data,
  muscle_name = NULL,
  joint_name = NULL,
  title_name = NULL
)
}
\arguments{
\item{joint_name}{joint name.}

\item{title_name}{title name.}

\item{Data}{Data.}
}
\description{
trajectory_color() can plot coordinates for trajectory analysis.
}
