% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_plot.R
\name{angle_plot}
\alias{angle_plot}
\title{plot for joint's angles}
\usage{
angle_plot(
  Data,
  sagittal = "left",
  LR = "left",
  file_name = NULL,
  angle_name = NULL
)
}
\arguments{
\item{sagittal}{left or right.}

\item{LR}{left or right.}

\item{file_name}{file name.}

\item{angle_name}{angle name.}

\item{data_array}{data array or data_list (it is obtained from function make_array()).}
}
\description{
plotting angles of a joint.
}
