% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch.R
\name{epoch}
\alias{epoch}
\title{epoch}
\usage{
epoch(data_array, maxtime)
}
\arguments{
\item{data_array}{data array.}

\item{maxtime}{maxtime.}
}
\description{
epoch convert data to list based on epoching by peak detection.
}
