% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif_skeleton.R
\name{gif_skeleton}
\alias{gif_skeleton}
\title{gif_skeleton}
\usage{
gif_skeleton(Array, interval = 0.2, filename)
}
\arguments{
\item{Array}{openpose joint coordinate array}

\item{interval}{interval parameter for gif animation}

\item{filename}{filepath}
}
\description{
gif_skeleton can visualize the skeleton of 1 frame
}
