% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_imp.R
\name{miss_imp}
\alias{miss_imp}
\title{miss_imp}
\usage{
miss_imp(data_array, thr = 0, lag = 5, percentile = 95, type = "linear")
}
\arguments{
\item{data_array}{data array.}

\item{thr}{threshold.}

\item{lag}{lag.}

\item{percentile}{percentile.}

\item{type}{"linear" or "spline".}
}
\description{
miss_imp() can impute missing data.
}
