% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resampling}
\alias{resampling}
\title{resampling}
\usage{
resampling(Array, original = NULL, to = NULL)
}
\arguments{
\item{Array}{array.}

\item{original}{original length}

\item{to}{arbitrary length}
}
\description{
convert array to arbitrary length array
}
