% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switching.R
\name{switching}
\alias{switching}
\title{switching}
\usage{
switching(epoch_list, lag = 1, thr = 0.3, type = "linear")
}
\arguments{
\item{epoch_list}{epoch list.}

\item{lag}{lag.}

\item{thr}{threshold.}

\item{type}{type "linear" or "spline".}
}
\description{
switching() can detect NA and impute them.
}
